review client page with chart

+++ src/main/java/com/user/planeapp/Flight.java
@@ -17,7 +17,7 @@ public class Flight {
         this.ticketSum = ticketSum;
     }
 
-    public double getFlightDuration() {
+    public int getFlightDuration() {
         return flightDuration;
     }
